synckolab 0.4.30
integrated all patches submitted by Andreas Gungl:
1) fix for a problem when a message has been removed from the server and
the message has already been in the internal hash database
(In this case the entry in the database hasn't been deleted which lead to
problems if the same events have been uploaded again to the server by
another client. Because then the events would be deleted again and again by
the first client.)

2) fix for dealing with more than one calendar sync settings
(some typos or copy & paste errors prevented the settings dialog to be updated
correctly)

3) allow parsing of most of the events generated by Toltec Connector
(the XML attachments in the message are base64-encoded, so the detection of
that encoding and the decoding of base64 is needed)

There are still minor problems with the calendar sync, but most of them are
hard to reproduce. If anyone has a reproducible test case, please let me
know.

- also added global variable for debug messages (so it can be configured later)